fix: resolve Rust compilation errors for CI/CD builds
- Fix type mismatch errors by casting i32 to u32 for PhysicalSize fields
- Add conditional compilation attribute for Command import to prevent unused import warning on non-Windows platforms

+++ rapitas-desktop/src-tauri/src/main.rs
@@ -290,8 +290,8 @@ async fn open_split_view(app: tauri::AppHandle, url: String) -> Result<(), Strin
             main_window
                 .set_size(tauri::Size::Physical(
                     tauri::PhysicalSize {
-                        width: screen_width / 2,
-                        height: screen_height,
+                        width: (screen_width / 2) as u32,
+                        height: screen_height as u32,
                     }
                 ))
                 .ok();
